// src/types/form.ts
export interface ReleaseFormValues {
    parentName: string;
    parentEmail: string;
    parentPhone: string;
    childName: string;
    childDOB: string;
    childAddress: string;
    childMedicalNotes: string;
    childDoctor: string;
    childInsurance: string;
    emergencyName: string;
    emergencyPhone: string;
    signature: string;
    dateSigned: string;
    liabilityText: string;
  }
  